seqCounter is now a separate TU
