+++ bronson.charles/P1/main.cpp
@@ -1,36 +1,11 @@
 #include <iostream>
-#include <limits>
 #include <stdexcept>
-
-struct SequenceCounter
-{
-public:
-  SequenceCounter():
-    seqLength_(0)
-  {}
-
-  void operator()(int number)
-  {
-     const size_t maxSize = std::numeric_limits< size_t >::max();
-     if (seqLength_ == maxSize)
-     {
-       throw std::logic_error("Sequence is too long.");
-     }
-     ++seqLength_;
-  }
-
-  size_t operator()() const
-  {
-    return seqLength_;
-  }
-
-private:
-  size_t seqLength_;
-};
+#include "seqCounter.hpp"
 
 int main()
 {
   int number = 0;
+  using namespace bronson;
   SequenceCounter sequenceCounter;
   do
   {
